+++ temba/flows/views.py
@@ -2129,11 +2129,17 @@ class FlowCRUDL(SmartCRUDL):
                         )
                     )
 
+                if self.flow.org.is_suspended:
+                    raise ValidationError(
+                        _(
+                            "Sorry, your account is currently suspended. "
+                            "To enable starting flows, please contact support."
+                        )
+                    )
                 if self.flow.org.is_flagged:
                     raise ValidationError(
                         _(
-                            "Sorry, your account is currently flagged. "
-                            "To enable sending messages, please contact support."
+                            "Sorry, your account is currently flagged. To enable starting flows, please contact support."
                         )
                     )