+++ temba/mailroom/modifiers.py
@@ -1,6 +1,11 @@
 from typing import Dict, List, NamedTuple
 
 
+class FieldRef(NamedTuple):
+    key: str
+    name: str
+
+
 class GroupRef(NamedTuple):
     uuid: str
     name: str
@@ -30,6 +35,17 @@ class Language(Modifier):
         self.language = language
 
 
+class Field(Modifier):
+    type = "field"
+
+    def __init__(self, field: FieldRef, value: str):
+        self.field = field
+        self.value = value
+
+    def as_def(self) -> Dict:
+        return {"type": self.type, "field": self.field._asdict(), "value": self.value}
+
+
 class Status(Modifier):
     type = "status"
 
@@ -40,9 +56,17 @@ class Status(Modifier):
 class Groups(Modifier):
     type = "groups"
 
-    def __init__(self, modification: str, groups: List[GroupRef]):
-        self.modification = modification
+    def __init__(self, groups: List[GroupRef], modification: str):
         self.groups = groups
+        self.modification = modification
 
     def as_def(self) -> Dict:
-        return {"type": self.type, "modification": self.modification, "groups": [g._asdict() for g in self.groups]}
+        return {"type": self.type, "groups": [g._asdict() for g in self.groups], "modification": self.modification}
+
+
+class URNs(Modifier):
+    type = "urns"
+
+    def __init__(self, urns: List[str], modification: str):
+        self.urns = urns
+        self.modification = modification