+++ temba/mailroom/tests.py
@@ -88,11 +88,13 @@ class MailroomClientTest(TembaTest):
                 [
                     modifiers.Name(name="Bob"),
                     modifiers.Language(language="fra"),
+                    modifiers.Field(field=modifiers.FieldRef(key="age", name="Age"), value="43"),
                     modifiers.Status(status="blocked"),
                     modifiers.Groups(
-                        modification="add",
                         groups=[modifiers.GroupRef(uuid="c153e265-f7c9-4539-9dbc-9b358714b638", name="Doctors")],
+                        modification="add",
                     ),
+                    modifiers.URNs(urns=["+tel+1234567890"], modification="append"),
                 ],
             )
             self.assertEqual("6393abc0-283d-4c9b-a1b3-641a035c34bf", response["1"]["contact"]["uuid"])
@@ -106,12 +108,14 @@ class MailroomClientTest(TembaTest):
                     "modifiers": [
                         {"type": "name", "name": "Bob"},
                         {"type": "language", "language": "fra"},
+                        {"type": "field", "field": {"key": "age", "name": "Age"}, "value": "43"},
                         {"type": "status", "status": "blocked"},
                         {
                             "type": "groups",
-                            "modification": "add",
                             "groups": [{"uuid": "c153e265-f7c9-4539-9dbc-9b358714b638", "name": "Doctors"}],
+                            "modification": "add",
                         },
+                        {"type": "urns", "urns": ["+tel+1234567890"], "modification": "append"},
                     ],
                 },
             )